refactor: use  kebab case naming for file
